[refactor]: useClickOutside custom hook 정의 - 컴포넌트 영역 외 클릭 시 특정 메소드 수행
